% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmapack-package.R
\docType{package}
\name{varmapack-package}
\alias{varmapack}
\alias{varmapack-package}
\title{Varmapack: Exact-Start Simulation of Gaussian Vector ARMA Models}
\description{
An R interface to the Varmapack C library for simulation and analysis of
Gaussian VAR, VMA, VARMA, and VARMAX models.
}
\seealso{
\itemize{
\item \code{vignette("getting-started", package = "varmapack")} for a tutorial.
\item \code{vignette("mathematical-description", package = "varmapack")} for the
mathematical description.
\item \href{https://github.com/jonasson2/varmapack}{Varmapack on GitHub}.
\item \href{https://github.com/jonasson2/varmapack/issues}{Report bugs}.
}
}
\author{
\strong{Maintainer}: Kristján Jónasson \email{jonasson@hi.is}

}
\keyword{internal}
