% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{varmapack_testcases}
\alias{varmapack_testcases}
\title{List named VARMA testcases}
\usage{
varmapack_testcases()
}
\value{
A data frame with columns \code{index}, \code{name}, \code{p}, \code{q}, and \code{r}.
}
\description{
List named VARMA testcases
}
\examples{
varmapack_testcases()

}
