% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmapack-package.R
\docType{package}
\name{varmapack-package}
\alias{varmapack}
\alias{varmapack-package}
\title{Varmapack: Exact-Start Simulation of Gaussian VARMA Models}
\description{
An R interface to the Varmapack C library for simulation and analysis of
Gaussian VAR, VMA, VARMA, and VARMAX models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonasson2/varmapack}
  \item Report bugs at \url{https://github.com/jonasson2/varmapack/issues}
}

}
\author{
\strong{Maintainer}: Kristján Jónasson \email{jonasson@hi.is}

}
\keyword{internal}
